Extract common DropGuard struct and NumberExt trait
Both are used in multiple schemars versions so keeping them in utils.rs
is a better choice than duplicating them in each file.
